Split files; db.go was getting really long
