+++ cluster/raft/logdeck/db.go
@@ -6,108 +6,7 @@ import (
 	"iter"
 	"os"
 	"path/filepath"
-	"regexp"
-	"slices"
-	"strconv"
-	"syscall"
-
-	"golang.org/x/exp/mmap"
 )
-
-type (
-	// DB represents a sequential collection of values, organized into Logs,
-	// and indexed by type and sequence. New values are appended to the active Log.
-	// Once the active Log reaches MaxLogSize, a new Log is provisioned to
-	// receive new appends. The old Log becomes read-only. Once DB exceeds
-	// MaxLogCount, compaction removes the oldest Log and its values.
-	DB interface {
-		// Append writes a value to the DB.
-		Append(t Type, value []byte) error
-		// Get retrieves a value with Type t at index i.
-		Get(t Type, i int) ([]byte, error)
-		// Count returns how many values with Type t are in DB.
-		Count(t Type) int
-		// First returns the first value of Type t in the DB. The value returned
-		// by this method changes after a compaction.
-		First(t Type) ([]byte, error)
-		// Last returns the last value of Type t that was written to the DB.
-		Last(t Type) ([]byte, error)
-		// Range returns an iterator that ranges over all values of Type t
-		// in the range [lo, hi[.
-		Range(t Type, lo, hi int) iter.Seq2[[]byte, error]
-		// Cut manually cuts a new Log in the DB. Cutting a Log is normally
-		// triggered automatically when MaxLogSize or MaxLogRecordCount is
-		// exceeded. Instead, Cut may be used to trigger them manually when more
-		// control is required, like for tests. As such, Cut does not consider
-		// MaxLogSize and MaxLogRecordCount.
-		Cut() error
-		// Compact manually triggers a compaction. Compactions are normally
-		// triggered automatically when MaxLogCount is exceeded. Instead,
-		// Compact may be used to trigger them manually when more control
-		// is required, like for tests. As such, Compact does not consider MaxLogCount.
-		// Attempting to Compact when the DB only contains one Log returns ErrInvalidCompaction.
-		Compact() error
-		// CutHook registers CutHookFunc f, which is run whenever a Log is cut.
-		// To clear the CutHook, call CutHook with a nil argument.
-		CutHook(f CutHookFunc)
-		// CompactHook registers CompactHook f, which is run whenever DB compacts a Log.
-		// To clear the CompactHook, call CompactHook with a nil argument.
-		CompactHook(f CompactHookFunc)
-		// Sync calls syscall.Fdatasync on the active Log, ensuring that buffered
-		// writes to it are flushed to disk. DB only syncs automatically when a Log
-		// is cut and becomes read-noly. Any more syncs are the application's responsibility.
-		Sync() error
-		// Close closes the DB, flushing any pending writes to disk.
-		Close() error
-	}
-
-	// Type represents the type of a value appended to the DB. Consumers of DB
-	// are expected to define their own types as constants of Type.
-	Type uint32
-
-	// LogID represents the sequential ID of a Log in the DB.
-	LogID uint64
-
-	// CutHookFunc is executed whenever a Log is cut. A Log is cut when it reaches
-	// its maximum size and is moved into read-only mode. A new Log is then
-	// provisioned to receive new writes. CutHookFunc is executed right after
-	// the new Log is provisioned. If Append is called on the LogDeck in
-	// CutHookFunc, the appended value is guaranteed to be the first in the new Log.
-	CutHookFunc func(id LogID) error
-
-	// CompactHookFunc is executed whenever a Log is compacted. Compaction is
-	// triggered when a newly provisioned Log causes MaxLogCount to be exceeded.
-	// CompactionHookFunc is executed right before the oldest Log is actually
-	// removed from DB, meaning that its data can still be queried for the
-	// duration of CompactHookFunc.
-	// TODO: Instead of Counters, maybe have it be an iter.Seq2[Type, uint64],
-	// the signature of Counters.All(). That way we can make Counters private.
-	CompactHookFunc func(c Counters) error
-
-	// Options defines the configurable options of the DB.
-	Options struct {
-		// MaxLogSize determines the maximum size that a single Log in the DB can have.
-		// When appending a value to a Log would make it grow larger than MaxLogSize,
-		// a new Log is provisioned, and the value is appended to the new Log.
-		// The total maximum DB size on disk is MaxLogSize * (MaxLogCount + 1).
-		MaxLogSize int64
-		// MaxLogValueCount determines the maximum amount of values that a single Log in the DB can have.
-		// When appending a value to a Log would make it exceed MaxLogValueCount,
-		// a new Log is provisioned, and the value is appended to the new Log.
-		MaxLogValueCount int64
-		// MaxLogCount determines how many Logs can be contained in the DB.
-		// Once exceeded, the oldest Log in the DB is compacted.
-		// The total maximum DB size on disk is MaxLogSize * (MaxLogCount + 1).
-		MaxLogCount int
-	}
-)
-
-// DefaultOptions defines the default Options values.
-var DefaultOptions = &Options{
-	MaxLogSize:       64 << 20,
-	MaxLogValueCount: 10000,
-	MaxLogCount:      16,
-}
 
 // Open intializes a DB in directory dir. If the directory already contains
 // Logs belonging to a DB, the DB reads the Logs to restore its in-memory state.
@@ -218,18 +117,18 @@ type db struct {
 	// If a Record will cause a Log to exceed its maximum allowed size,
 	// a new Log will be provisioned and the Record will be appended there.
 	maxLogRecordCount int64
-	// maxLogCount determines the maximum amount of logs stored in the Deck.
+	// maxLogCount determines the maximum amount of logs stored in the DB.
 	// When the number of logs exceeds this number, the oldest log will be compacted.
 	maxLogCount int
-	// Inventory holds pointers to all known Records in the Deck,
-	// organized by type. It grows when appending Records to the Deck,
-	// and shrinks when Logs in the Deck are compacted.
+	// Inventory holds pointers to all known Records in the DB,
+	// organized by type. It grows when appending Records to the DB,
+	// and shrinks when Logs in the DB are compacted.
 	inventory *inventory
-	// cutHook is provided by the Deck consumer. If provided,
-	// it is called by Deck after every Log is cut.
+	// cutHook is provided by the DB consumer. If provided,
+	// it is called by DB after every Log is cut.
 	cutHook CutHookFunc
-	// compactHook is provided by the Deck consumer. If provided,
-	// it is called by Deck after every compaction, allowing the consumer
+	// compactHook is provided by the DB consumer. If provided,
+	// it is called by DB after every compaction, allowing the consumer
 	// to update its internal bookkeeping.
 	compactHook CompactHookFunc
 }
@@ -257,7 +156,7 @@ func (d *db) Append(t Type, value []byte) error {
 	d.inventory.Add(r.Type, d.pointer())
 
 	// Run compaction _after_ adding the new entry so that
-	// the compaction handler has an up-to-date view of the Deck.
+	// the compaction handler has an up-to-date view of the DB.
 	if len(d.logs) > d.maxLogCount {
 		return d.compact()
 	}
@@ -404,7 +303,7 @@ func (d *db) compact() error {
 	log := d.logs[0]
 
 	// The CompactHook is run _before_ compaction actually occurs
-	// so that the Deck consumer has a full view of the data, including
+	// so that the DB consumer has a full view of the data, including
 	// what is being removed. An errors encountered by compaction are surfaced
 	// to the application.
 	if d.compactHook != nil {
@@ -440,114 +339,4 @@ func (d *db) pointer() pointer {
 		Offset: offset,
 		Size:   size,
 	}
-}
-
-func createLogFile(dir string, id LogID, size int64) (*logFile, error) {
-	filename := filepath.Join(dir, fmt.Sprintf("%020d.log", id))
-
-	w, err := os.OpenFile(filename, os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		return nil, err
-	}
-	err = syscall.Fallocate(int(w.Fd()), 0, 0, size)
-	if err != nil {
-		return nil, err
-	}
-
-	r, err := mmap.Open(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	return &logFile{
-		ID:   id,
-		File: w,
-		Mmap: r,
-		log:  newLog(r, w),
-	}, nil
-}
-
-func openLogFile(filename string) (*logFile, error) {
-	basename := filepath.Base(filename)
-	id, err := strconv.ParseUint(basename[:len(basename)-4], 10, 64)
-	if err != nil {
-		return nil, err
-	}
-
-	w, err := os.OpenFile(filename, os.O_WRONLY, 0644)
-	if err != nil {
-		return nil, err
-	}
-
-	r, err := mmap.Open(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	return &logFile{
-		ID:   LogID(id),
-		File: w,
-		Mmap: r,
-		log:  newLog(r, w),
-	}, nil
-}
-
-// logFile represents a Log that is backed by a file.
-// It wraps the basic Log and adds an ID for tracking unique Logs,
-// and the handles of the underlying file.
-type logFile struct {
-	*log
-	ID   LogID
-	File *os.File
-	Mmap *mmap.ReaderAt
-}
-
-func (l *logFile) Sync() error {
-	return syscall.Fdatasync(int(l.File.Fd()))
-}
-
-func (l *logFile) Lock() error {
-	if err := l.Sync(); err != nil {
-		return err
-	}
-
-	if err := os.Truncate(l.File.Name(), l.cursor); err != nil {
-		return err
-	}
-
-	return l.File.Close()
-}
-
-func (l *logFile) Close() error {
-	if err := l.File.Close(); err != nil {
-		if !errors.Is(err, os.ErrClosed) {
-			return err
-		}
-	}
-	return l.Mmap.Close()
-}
-
-var logNameRe = regexp.MustCompile(`(\d{20}).log`)
-
-// discoverLogFiles searches for files with names matching logNameRe in directory dir,
-// returning a sorted list of filenames.
-func discoverLogFiles(dir string) ([]string, error) {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		return nil, err
-	}
-
-	logFiles := make([]string, 0)
-	for _, e := range entries {
-		if e.IsDir() {
-			continue
-		}
-
-		if logNameRe.MatchString(e.Name()) {
-			logFiles = append(logFiles, e.Name())
-		}
-	}
-
-	slices.Sort(logFiles)
-	return logFiles, nil
 }
